test: SystemInterface covered by a single test module
The case-only duplicate pair test_SystemInterface.py /
test_system_interface.py tested the same class with near-identical
test names and collided on case-insensitive filesystems. The
lowercase module survives with the CamelCase file's unique cases
(from_system construction, kwargs merged into a provided updates
dict, index-only merge_variable_inputs path) ported in.
